Fix FormatMessage definition for x64
 - FFI :ulong type is platform dependent, and FormatMessage expects
   DWORD parameters, which are always 32-bit wide unsigned integers
   regardless of platform

+++ lib/win32/windows/helper.rb
@@ -7,7 +7,7 @@ module Windows
     ffi_lib :kernel32
 
     attach_function :FormatMessage, :FormatMessageA,
-    [:ulong, :pointer, :ulong, :ulong, :pointer, :ulong, :pointer], :ulong
+    [:uint32, :pointer, :uint32, :uint32, :pointer, :uint32, :pointer], :uint32
 
     def win_error(function, err=FFI.errno)
       flags = 0x00001000 | 0x00000200
